requirements.txt
argon2-cffi==18.3.0
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.11.29
cffi==1.11.5
cryptography==2.4.1
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
django-ipware==1.0.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.7
passlib==1.7.1
psycopg2==2.8.2
psycopg2-binary==2.8.2
pycparser==2.19
pycrypto==2.6.1
pygeoip==0.3.2
Pygments==2.3.1
pytz==2018.7
simple-crypt==4.1.7
six==1.11.0
whitenoise==4.1.2
